Convert kmeans module to package
