+++ spdivik/kmeans/__init__.py
@@ -1,0 +1,4 @@
+"""Numpy-based implementation of k-means algorithm"""
+
+from ._initialization import ExtremeInitialization, Initialization
+from ._core import KMeans, Labeling
