+++ test/kmeans/__init__.py
@@ -1,0 +1,10 @@
+import numpy as np
+
+data = np.array([
+    [1, 1, 1, 1],
+    [2, 4, 2, 2],
+    [1.9, 4.2, 1.9, 1.9],
+    [2, 2, 2, 2],
+    [1.1, 0.8, 1.1, 1.1],
+    [1000, 490231, -412342, -7012]
+])
